+++ tortoise/models.py
@@ -1,9 +1,11 @@
 import operator
+from typing import Dict, Optional, Set  # noqa
 
 from pypika import Table, functions
 from pypika.enums import SqlTypes
 
 from tortoise import fields
+from tortoise.backends.base.client import BaseDBAsyncClient  # noqa
 from tortoise.exceptions import ConfigurationError, OperationalError
 from tortoise.fields import ManyToManyRelationManager, RelationQueryContainer
 from tortoise.queryset import QuerySet
@@ -129,7 +131,7 @@ def get_backward_fk_filters(field_name, field):
     return filters
 
 
-def get_filters_for_field(field_name: str, field: fields.Field, source_field: str):
+def get_filters_for_field(field_name: str, field: Optional[fields.Field], source_field: str):
     if isinstance(field, fields.ManyToManyField):
         return get_m2m_filters(field_name, field)
     filters = {
@@ -204,31 +206,34 @@ def get_filters_for_field(field_name: str, field: fields.Field, source_field: st
 
 
 class MetaInfo:
+    __slots__ = ('abstract', 'table', 'app', 'fields', 'db_fields', 'm2m_fields', 'fk_fields',
+                 'backward_fk_fields', 'fetch_fields', 'fields_db_projection',
+                 'fields_db_projection_reverse', 'filters', 'fields_map', 'db')
+
     def __init__(self, meta):
-        self.abstract = getattr(meta, 'abstract', False)
-        self.table = getattr(meta, 'table', None)
-        self.app = getattr(meta, 'app', 'models')
-        self.fields = set()
-        self.db_fields = set()
-        self.m2m_fields = set()
-        self.fk_fields = set()
-        self.backward_fk_fields = set()
-        self.fetch_fields = set()
-        self.fields_db_projection = {}
-        self.fields_db_projection_reverse = {}
-        self.filters = {}
-        self.fields_map = {}
-        self.db = None
+        self.abstract = getattr(meta, 'abstract', False)  # type: bool
+        self.table = getattr(meta, 'table', None)  # type: Optional[Table]
+        self.app = getattr(meta, 'app', 'models')  # type: str
+        self.fields = set()  # type: Set[str]
+        self.db_fields = set()  # type: Set[str]
+        self.m2m_fields = set()  # type: Set[str]
+        self.fk_fields = set()  # type: Set[str]
+        self.backward_fk_fields = set()  # type: Set[str]
+        self.fetch_fields = set()  # type: Set[str]
+        self.fields_db_projection = {}  # type: Dict[str,str]
+        self.fields_db_projection_reverse = {}  # type: Dict[str,str]
+        self.filters = {}  # type: Dict[str, Dict[str, dict]]
+        self.fields_map = {}  # type: Dict[str, fields.Field]
+        self.db = None  # type: Optional[BaseDBAsyncClient]
 
 
 class ModelMeta(type):
     def __new__(mcs, name, bases, attrs, *args, **kwargs):
-        fields_db_projection = {}
-        fields_map = {}
-        filters = {}
-        fk_fields = set()
-        m2m_fields = set()
-        meta = MetaInfo(attrs.get('Meta'))
+        fields_db_projection = {}  # type: Dict[str,str]
+        fields_map = {}  # type: Dict[str, fields.Field]
+        filters = {}  # type: Dict[str, Dict[str, dict]]
+        fk_fields = set()  # type: Set[str]
+        m2m_fields = set()  # type: Set[str]
 
         for key, value in attrs.items():
             if isinstance(value, fields.Field):
@@ -261,28 +266,34 @@ class ModelMeta(type):
                             source_field=fields_db_projection[key]
                         )
                     )
-        new_class = super().__new__(mcs, name, bases, attrs)
 
-        new_class._meta = meta
-        new_class._meta.fields_map = fields_map
-        new_class._meta.fields_db_projection = fields_db_projection
-        new_class._meta.fields_db_projection_reverse = {
+        attrs['_meta'] = meta = MetaInfo(attrs.get('Meta'))
+
+        if 'id' not in attrs:
+            attrs['id'] = None
+
+        meta = meta
+        meta.fields_map = fields_map
+        meta.fields_db_projection = fields_db_projection
+        meta.fields_db_projection_reverse = {
             value: key
             for key, value in fields_db_projection.items()
         }
-        new_class._meta.fields = set(fields_map.keys())
-        new_class._meta.db_fields = set(fields_db_projection.values())
-        new_class._meta.filters = filters
-        new_class._meta.fk_fields = fk_fields
-        new_class._meta.backward_fk_fields = set()
-        new_class._meta.m2m_fields = m2m_fields
-        new_class._meta.fetch_fields = fk_fields | m2m_fields
-        new_class._meta.db = None
+        meta.fields = set(fields_map.keys())
+        meta.db_fields = set(fields_db_projection.values())
+        meta.filters = filters
+        meta.fk_fields = fk_fields
+        meta.backward_fk_fields = set()
+        meta.m2m_fields = m2m_fields
+        meta.fetch_fields = fk_fields | m2m_fields
+        meta.db = None
         if not fields_map:
-            new_class._meta.abstract = True
-        if not new_class._meta.abstract:
+            meta.abstract = True
+
+        new_class = super().__new__(mcs, name, bases, attrs)
+        if not meta.abstract:
             from tortoise import Tortoise
-            Tortoise.register_model(new_class._meta.app, new_class.__name__, new_class)
+            Tortoise.register_model(meta.app, new_class.__name__, new_class)
         return new_class
 
 
@@ -380,7 +391,7 @@ class Model(metaclass=ModelMeta):
         return '<{}: {}>'.format(self.__class__.__name__, self.__str__())
 
     def __hash__(self):
-        if not hasattr(self.id) or not self.id:
+        if self.id:
             raise TypeError('Model instances without id are unhashable')
         return hash(self.id)
 
@@ -419,8 +430,7 @@ class Model(metaclass=ModelMeta):
 
     @classmethod
     def get(cls, *args, **kwargs):
-        # TODO: Raise exception if no result
-        return QuerySet(cls).filter(*args, **kwargs).first()
+        return QuerySet(cls).get(*args, **kwargs)
 
     @classmethod
     async def fetch_for_list(cls, instance_list, *args, using_db=None):