+++ tortoise/tests/testmodels.py
@@ -131,3 +131,17 @@ class JSONFields(Model):
     id = fields.IntField(pk=True)
     data = fields.JSONField()
     data_null = fields.JSONField(null=True)
+
+
+# TODO: Test that minimaly specified relations work as expected
+class MinRelation(Model):
+    id = fields.IntField(pk=True)
+    tournament = fields.ForeignKeyField('models.Tournament')
+    participants = fields.ManyToManyField(
+        'models.Team'
+    )
+
+
+# TODO: Test that no id models work as expected
+class NoID(Model):
+    char = fields.CharField(max_length=255, null=True)