+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.9.4
+-----
+- No more asserts, only Tortoise Exceptions
+- Fixed PyLint plugin to work with pylint>=2.0.0
+- Formalised unittest classes & documented them.
+- ``__slots__`` where it was easy to do. (Changes class instances from dicts into tuples, memory savings)
+
 0.9.3
 -----
 - Fixed backward incompatibility for Python 3.7
@@ -22,7 +29,7 @@ Changelog
 -----
 - Added support for nested queries for ``values`` and ``values_list``:
 
-  .. code-block:: python3
+.. code-block:: python3
 
     result = await Event.filter(id=event.id).values('id', 'name', tournament='tournament__name')
     result = await Event.filter(id=event.id).values_list('id', 'participants__name')
@@ -56,7 +63,7 @@ Changelog
 -----
 - Added schema generation and more options for fields:
 
-  .. code-block:: python3
+.. code-block:: python3
 
     from tortoise import Tortoise
     from tortoise.backends.sqlite.client import SqliteClient
@@ -71,7 +78,7 @@ Changelog
 -----
 - Added filtering and ordering by related models fields:
 
-  .. code-block:: python3
+.. code-block:: python3
 
     await Tournament.filter(
         events__name__in=['1', '3']